+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/graph/DependencyGraphNodeCache.kt
@@ -29,6 +29,7 @@ import dev.zacsweers.metro.compiler.ir.annotationsIn
 import dev.zacsweers.metro.compiler.ir.bindingContainerClasses
 import dev.zacsweers.metro.compiler.ir.createIrBuilder
 import dev.zacsweers.metro.compiler.ir.excludedClasses
+import dev.zacsweers.metro.compiler.ir.findInjectableConstructor
 import dev.zacsweers.metro.compiler.ir.isAccessorCandidate
 import dev.zacsweers.metro.compiler.ir.isAnnotatedWithAny
 import dev.zacsweers.metro.compiler.ir.isBindingContainer
@@ -77,10 +78,12 @@ import org.jetbrains.kotlin.ir.expressions.IrConstructorCall
 import org.jetbrains.kotlin.ir.expressions.IrExpression
 import org.jetbrains.kotlin.ir.expressions.IrVararg
 import org.jetbrains.kotlin.ir.symbols.IrSymbol
+import org.jetbrains.kotlin.ir.types.classFqName
 import org.jetbrains.kotlin.ir.types.classOrFail
 import org.jetbrains.kotlin.ir.types.classOrNull
 import org.jetbrains.kotlin.ir.types.isUnit
 import org.jetbrains.kotlin.ir.types.typeWith
+import org.jetbrains.kotlin.ir.util.callableId
 import org.jetbrains.kotlin.ir.util.classId
 import org.jetbrains.kotlin.ir.util.classIdOrFail
 import org.jetbrains.kotlin.ir.util.defaultType
@@ -177,6 +180,7 @@ internal class DependencyGraphNodeCache(
       (metroGraph ?: graphDeclaration).allSupertypesSequence(excludeSelf = false).toList()
 
     private var hasGraphExtensions = false
+    private var hasErrors = false
 
     private fun computeDeclaredScopes(): Set<IrAnnotation> {
       return buildSet {
@@ -350,6 +354,8 @@ internal class DependencyGraphNodeCache(
           MetroDiagnostics.GRAPH_DEPENDENCY_CYCLE,
           message,
         )
+        // In this case, we exit early as we have a self-cycle in the graph that deferring would
+        // just loop
         exitProcessing()
       }
     }
@@ -389,6 +395,7 @@ internal class DependencyGraphNodeCache(
         MetroDiagnostics.METRO_ERROR,
         message,
       )
+      hasErrors = true
     }
 
     fun build(): DependencyGraphNode {
@@ -620,7 +627,37 @@ internal class DependencyGraphNodeCache(
                 contextKey.typeKey.copy(contextKey.typeKey.type.wrapInMembersInjector())
               val finalContextKey = contextKey.withTypeKey(memberInjectorTypeKey)
 
-              injectors += InjectorFunction(finalContextKey, metroFunction)
+              // Check if the target is constructor-injected. We need to do this in IR too because
+              // FIR will miss inherited injectors. https://github.com/ZacSweers/metro/issues/1606
+              val hasInjectConstructor =
+                declaration.isFakeOverride &&
+                  declaration.regularParameters[0]
+                    .type
+                    .rawTypeOrNull()
+                    ?.findInjectableConstructor(false) != null
+
+              if (hasInjectConstructor) {
+                // If the original declaration is in our compilation, report it. Otherwise fall
+                // through to the nearest available declaration to report.
+                val originalDeclaration = declaration.overriddenSymbolsSequence().last().owner
+                val isExternal = originalDeclaration.isExternalParent
+                val middle =
+                  if (isExternal) {
+                    val callableId = originalDeclaration.callableId
+                    // It's an external declaration, so make it clear which function
+                    "the inherited '${callableId.callableName}' inject function from '${callableId.classId!!.asFqNameString()}'"
+                  } else {
+                    "this inject function"
+                  }
+                metroContext.reportCompat(
+                  originalDeclaration.takeUnless { isExternal } ?: declaration,
+                  MetroDiagnostics.SUSPICIOUS_MEMBER_INJECT_FUNCTION,
+                  "Injected class '${declaration.regularParameters[0].type.classFqName!!.asString()}' is constructor-injected and can be instantiated by Metro directly, so $middle is unnecessary.",
+                )
+                hasErrors = true
+              } else {
+                injectors += InjectorFunction(finalContextKey, metroFunction)
+              }
             } else {
               // Accessor or binds
               val metroFunction = metroFunctionOf(declaration, annotations)
@@ -977,6 +1014,11 @@ internal class DependencyGraphNodeCache(
             }
           },
         )
+        exitProcessing()
+      }
+
+      // Exit after collecting all errors
+      if (hasErrors) {
         exitProcessing()
       }